+++ src/app/components/file-distribution/chunks-visualizer.js
@@ -41,7 +41,7 @@ export default Component.extend(I18n, {
    * @virtual
    * @type {number}
    */
-  blockCount: undefined,
+  chunkCount: undefined,
 
   /**
    * @virtual
@@ -90,15 +90,15 @@ export default Component.extend(I18n, {
   /**
    * @type {Ember.ComputedProperty<string>}
    */
-  blocksSizeText: computed('fileSizeOnStorage', 'blockCount', function blocksSizeText() {
+  chunksSizeText: computed('fileSizeOnStorage', 'chunkCount', function chunksSizeText() {
     const {
       fileSizeOnStorage,
-      blockCount,
-    } = this.getProperties('fileSizeOnStorage', 'blockCount');
-    return this.t('blocksSize', {
+      chunkCount,
+    } = this;
+    return this.t('chunksSize', {
       size: bytesToString(fileSizeOnStorage),
-      blockCount,
-      blockNoun: blockCount === 1 ? this.t('block') : this.t('blocks'),
+      chunkCount,
+      chunkNoun: chunkCount === 1 ? this.t('chunk') : this.t('chunks'),
     });
   }),